+++ SEAssignment/PaymentPoints.cs
@@ -3,7 +3,7 @@
 public class PaymentPoints : Payment
 {
 
-	public PaymentPoints(Receipt r, string purp, decimal amt) : base(r, purp, amt) { }
+	public PaymentPoints(Receipt r, string purp, double amt) : base(r, purp, amt) { }
 
 	public override void pay() { }
